Issue 306 - code split rpc
The Communication Layer is handled by the rpc.ts file, where the main endpoints
to use the IPA API commands are defined.  IPA API currently has more than 300
commands, so keeping them in a single file can lead to a humongous file making
it difficult to maintain.

We can use the RTK Query's code injection feature to split the endpoints into
different files and inject them to the main API.

Fixes: https://github.com/freeipa/freeipa-webui/issues/306
