+++ src/pages/PreservedUsers/PreservedUsers.tsx
@@ -42,11 +42,8 @@ import { FetchBaseQueryError } from "@reduxjs/toolkit/dist/query";
 import { SerializedError } from "@reduxjs/toolkit";
 // Utils
 import { API_VERSION_BACKUP, isUserSelectable } from "src/utils/utils";
-import {
-  useGettingPreservedUserQuery,
-  GenericPayload,
-  useSearchEntriesMutation,
-} from "../../services/rpc";
+import { GenericPayload, useSearchEntriesMutation } from "../../services/rpc";
+import { useGettingPreservedUserQuery } from "../../services/rpcUsers";
 import useApiError from "src/hooks/useApiError";
 import GlobalErrors from "src/components/errors/GlobalErrors";
 import ModalErrors from "src/components/errors/ModalErrors";
